Replace ssize_t with int to keep msvc2019 happy
ssize_t results in compilation warnings. This code is temporary
regardless, so just replace with int to be done with them

+++ libvisca.c
@@ -105,7 +105,7 @@ _VISCA_loop(VISCAInterface_t *iface)
 {
 	VISCAPacket_t *pkt = &iface->ipacket;
 	unsigned char buffer[64];
-	ssize_t i, count;
+	int i, count;
 
 	count = _VISCA_read_bytes(iface, buffer, sizeof(buffer));
 	if (count <= 0)
